+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/presenters/YTSignInPresenter.java
@@ -7,7 +7,7 @@ import com.liskovsoft.mediaserviceinterfaces.ServiceManager;
 import com.liskovsoft.sharedutils.mylogger.Log;
 import com.liskovsoft.sharedutils.rx.RxHelper;
 import com.liskovsoft.smartyoutubetv2.common.app.presenters.dialogs.AccountSelectionPresenter;
-import com.liskovsoft.youtubeapi.service.YouTubeServiceManager;
+import com.liskovsoft.smartyoutubetv2.common.misc.MediaSourceRegistry;
 
 import io.reactivex.disposables.Disposable;
 
@@ -24,7 +24,7 @@ public class YTSignInPresenter extends SignInPresenter {
 
     private YTSignInPresenter(Context context) {
         super(context);
-        mService = YouTubeServiceManager.instance();
+        mService = MediaSourceRegistry.getServiceManager();
     }
 
     public static YTSignInPresenter instance(Context context) {