+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/presenters/dialogs/menu/providers/channelgroup/ChannelGroupServiceWrapper.java
@@ -9,7 +9,7 @@ import com.liskovsoft.mediaserviceinterfaces.data.ItemGroup;
 import com.liskovsoft.mediaserviceinterfaces.data.ItemGroup.Item;
 import com.liskovsoft.smartyoutubetv2.common.prefs.AppPrefs;
 import com.liskovsoft.smartyoutubetv2.common.prefs.AppPrefs.ProfileChangeListener;
-import com.liskovsoft.youtubeapi.service.YouTubeServiceManager;
+import com.liskovsoft.smartyoutubetv2.common.misc.MediaSourceRegistry;
 
 import java.io.File;
 import java.util.List;
@@ -99,7 +99,7 @@ public class ChannelGroupServiceWrapper implements ProfileChangeListener {
 
     private ChannelGroupService getService() {
         if (mService == null) {
-            mService = YouTubeServiceManager.instance().getChannelGroupService();
+            mService = MediaSourceRegistry.getServiceManager().getChannelGroupService();
         }
 
         return mService;